requirements.txt
humanize
jinja2
pandas
pandera[pandas]
python-dateutil
quantmod # micropip.install: ValueError: Requested 'narwhals>=1.15.1', but narwhals==1.12.1 is already installed
shiny
shiny_validate
yfinance # micropip.install: ValueError: Can't find a pure Python 3 wheel for 'curl-cffi>=0.7'.

# testing
# pytest
# pytest-playwright